Inserted a cast to fix Pras' bad build.
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@1138 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ coupled/src/solver/cardiac/BidomainDg0Assembler.hpp
@@ -351,7 +351,7 @@ public:
         assert(fixedExtracellularPotentialNodes.size() > 0);
         for(unsigned i=0; i<fixedExtracellularPotentialNodes.size(); i++)
         {
-            assert( fixedExtracellularPotentialNodes[i] < mpMesh->GetNumNodes() );
+            assert( (int) fixedExtracellularPotentialNodes[i] < mpMesh->GetNumNodes() );
         }
         mFixedExtracellularPotentialNodes = fixedExtracellularPotentialNodes;
     }
